browser-service/requirements.txt
playwright
flask
python-dotenv
openai

